refactor(reposicao): extrai 4 seções de AdminReposicaoAumentoDetail (738→387)
Quebra o god-component em componentes presentacionais na pasta aumentoDetail/
(já existente com types/shared/CancelarButton/ItemRow). Movimento puro,
behavior-preserving — sem mudança de lógica:

- DetalhesTab: card de extração Vision + form "Dados do anúncio"
- CategoriasTab: tabela + loop de ItemRow
- SkusAfetadosTab: tabela de SKUs afetados (presentacional puro)
- EstadoAcoesSidebar: sidebar "Estado e ações" com gating do botão Ativar

Os 4 novos files entram no tsconfig.strict.json. Estado/queries/mutações
permanecem na página, que vira container puro.
