+++ tsconfig.strict.json
@@ -373,6 +373,10 @@
     "src/components/reposicao/negociacaoParalela/types.ts",
     "src/components/reposicao/negociacaoParalela/helpers.ts",
     "src/components/reposicao/negociacaoParalela/SugestaoCard.tsx",
+    "src/components/reposicao/aumentoDetail/DetalhesTab.tsx",
+    "src/components/reposicao/aumentoDetail/CategoriasTab.tsx",
+    "src/components/reposicao/aumentoDetail/SkusAfetadosTab.tsx",
+    "src/components/reposicao/aumentoDetail/EstadoAcoesSidebar.tsx",
     "src/pages/AdminMonthlyReports.tsx",
     "src/pages/NfeReceipt.tsx",
     "src/pages/ResetPassword.tsx",
